Произошло слияние
